refactor: external 패키지 infra 패키지 하위 레이어로 변경
