--- src/main/java/com/app/groupdeal/external/pg/PaymentGatewayResponseDto.java
@@ -1,4 +1,0 @@
-package com.app.groupdeal.external.pg;
-
-public class PaymentGatewayResponseDto {
-}
