--- src/main/java/com/app/groupdeal/external/pg/PaymentGatewayService.java
@@ -1,4 +1,0 @@
-package com.app.groupdeal.external.pg;
-
-public interface PaymentGatewayService {
-}
