--- src/main/java/com/app/groupdeal/external/pg/toss/TossPaymentClient.java
@@ -1,4 +1,0 @@
-package com.app.groupdeal.external.pg.toss;
-
-public interface TossPaymentClient {
-}
