--- src/main/java/com/app/groupdeal/external/pg/toss/TossPaymentDto.java
@@ -1,4 +1,0 @@
-package com.app.groupdeal.external.pg.toss;
-
-public class TossPaymentDto {
-}
