+++ src/main/java/com/app/groupdeal/infrastructure/client/pg/PaymentGatewayMediator.java
@@ -1,0 +1,4 @@
+package com.app.groupdeal.infrastructure.client.pg;
+
+public class PaymentGatewayMediator {
+}
