+++ src/main/java/com/app/groupdeal/infrastructure/client/pg/PaymentGatewayResponseDto.java
@@ -1,0 +1,4 @@
+package com.app.groupdeal.infrastructure.client.pg;
+
+public class PaymentGatewayResponseDto {
+}
