+++ src/main/java/com/app/groupdeal/infrastructure/client/pg/PaymentGatewayService.java
@@ -1,0 +1,4 @@
+package com.app.groupdeal.infrastructure.client.pg;
+
+public interface PaymentGatewayService {
+}
