+++ src/main/java/com/app/groupdeal/infrastructure/client/pg/toss/TossPaymentClient.java
@@ -1,0 +1,4 @@
+package com.app.groupdeal.infrastructure.client.pg.toss;
+
+public interface TossPaymentClient {
+}
