+++ src/main/java/com/app/groupdeal/infrastructure/client/pg/toss/TossPaymentServiceImpl.java
@@ -1,0 +1,4 @@
+package com.app.groupdeal.infrastructure.client.pg.toss;
+
+public class TossPaymentServiceImpl {
+}
